requirements.txt
google-generativeai
python-dotenv
pydantic
supabase
python-fasthtml
monsterUi
pydantic